Playwright is not threadsafe. (see https://playwright.dev/python/docs/intro#threading)
By creating the Playwright instance as a global variable,
it introduces a mess of conflicts with Playwright and its backend loop

This change ensures that the playwright instance created is thread-local, and exposes a `shutdown()` method to safely clean up the TikTokAPI instance.

The API instance can now be used as a context manager, guaranteeing safe cleanup of the instance.

```py
with TikTokApi() as api:
    for trending_video in api.trending.videos():
        # do something
```

Includes other maintenance items that should be self explanitory.

Signed-off-by: Montel Edwards <m@monteledwards.com>